fixing clippy warnings, bumping version

+++ src/sqrt_price_math.rs
@@ -255,7 +255,7 @@ pub fn get_amount_0_delta(
         Ok(-I256::from_raw(_get_amount_0_delta(
             sqrt_ratio_b_x_96,
             sqrt_ratio_a_x_96,
-            -liquidity as i128,
+            -liquidity,
             false,
         )?))
     } else {
@@ -277,7 +277,7 @@ pub fn get_amount_1_delta(
         Ok(-I256::from_raw(_get_amount_1_delta(
             sqrt_ratio_b_x_96,
             sqrt_ratio_a_x_96,
-            -liquidity as i128,
+            -liquidity,
             false,
         )?))
     } else {
